requirements.txt
albumentations==2.0.8
numpy==2.3.1
pandas==2.3.0
Pillow==11.2.1
pytorch_lightning==2.5.1
segmentation_models_pytorch==0.5.0
torch==2.6.0+cu126
